[doorbird] Change relay/IR switch to button
