+++ homeassistant/components/doorbird/const.py
@@ -2,7 +2,7 @@
 from homeassistant.const import Platform
 
 DOMAIN = "doorbird"
-PLATFORMS = [Platform.SWITCH, Platform.CAMERA]
+PLATFORMS = [Platform.BUTTON, Platform.CAMERA]
 DOOR_STATION = "door_station"
 DOOR_STATION_INFO = "door_station_info"
 DOOR_STATION_EVENT_ENTITY_IDS = "door_station_event_entity_ids"
